feat: dont crash when rendering invalid offers and purse amounts
